Add files via upload
This is a foundational example illustrating the integration of QML (Qt Meta-Object Language) and the Model-View-Controller (MVC) architectural pattern. The key aspect highlighted here is the usage of the Q_PROPERTY macro, which facilitates seamless communication between QML, responsible for the user interface, and C++, responsible for the underlying application logic and data management.

In this example, the MVC components are represented by the StudentModel class, serving as the data model, and the StudentController class, responsible for handling the application's logic. Q_PROPERTY is employed within the StudentModel to expose its properties, such as 'name' and 'id,' making them accessible and bindable in QML.